pyproject.toml
[project]
name = "all-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "bs4>=0.0.2",
    "mock>=5.2.0",
    "openai-agents>=0.3.2",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
